requirements.txt
numpy==2.1.3
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
pdf2image==1.17.0
Pillow==11.0.0
pytesseract==0.3.13
ray==2.38.0
scipy==1.14.1
torch==2.5.1+cu121
transformers==4.46.2
